requirements.txt
absl-py==2.1.0
altair==5.3.0
archspec==0.2.3
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
boltons==23.0.0
Brotli==1.0.9
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
distro==1.9.0
filelock==3.13.1
firebase-admin==6.5.0
firebase-auth==3.0.28
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.2
fsspec==2024.6.1
gast==0.6.0
gcloud==0.17.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
h5py==3.11.0
httplib2==0.22.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jws==0.1.3
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
numpy==1.26.4
oauth2client==3.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==3.10.0
pluggy==1.0.0
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycosat==0.6.6
pycparser==2.21
pycryptodome==3.10.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-jwt==2.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.17.21
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
sympy==1.12
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
termcolor==2.4.0
tf_keras==2.17.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
truststore==0.8.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.74
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==1.26.19
watchdog==4.0.1
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
wrapt==1.16.0
zstandard==0.22.0